Only add space if variable declaration needs a type.  Correct spelling of traversal.
In Python, variables need no type declaration.  We can specify an empty string for
the declaration, but adding a space (between the "empty" string and the variable
name) invalidates Python spacing.  Add a check to only add the space if the type
declaration string is non-empty.

Correct spelling of traversal, both in comments and in operational code.