Declare local variables in the same order.  Factor code to create assign statements.
The place that "randomizes" variable declaration is in complexities_generator.py  The
code was
                for __, n in enumerate(list(local_var[t])):
enumerate() adds an index, which is immediately discarded! (The __ part.) This code
makes sure variable names are in a consistent order
                # declare and initialize each variables of this type
                for n in sorted(list(local_var[t])):

Factor out a common function that creates assignment statements (" = ").  Use it in
the two places that create assignment statements.
complexities_generator.py:
            local_var_code += n + " = " + init + sterm + "\n"
generator.py
            init_var = compl_gen.out_ext_name+" = "+compl_gen.in_ext_name+sterm
I call it make_assign and put it in generator.py.  Maybe a prefix, like "synth" or a
separate file (module) would be better.